// src/app/commun/enumerations/codes-aides.enum.ts
export enum CodesAidesEnum {
  AGEPI = 'AGEPI',
  AIDE_MOBILITE = 'AM',
  AIDES_LOGEMENT = 'AL',
  AIDE_PERSONNALISEE_LOGEMENT = 'APL',
  ALLOCATION_ADULTES_HANDICAPES = 'AAH',
  ALLOCATION_LOGEMENT_FAMILIALE = 'ALF',
  ALLOCATION_LOGEMENT_SOCIALE = 'ALS',
  AIDE_RETOUR_EMPLOI = 'ARE',
  ALLOCATION_SOLIDARITE_SPECIFIQUE = 'ASS',
  ALLOCATION_SOUTIEN_FAMILIAL = 'ASF',
  ALLOCATIONS_FAMILIALES = 'AF',
  COMPLEMENT_AIDE_RETOUR_EMPLOI = 'cARE',
  COMPLEMENT_FAMILIAL = 'CF',
  CRC = 'CRC',
  CRDS = 'CRDS',
  CSG = 'CSG',
  IMMOBILIER = 'IMMO',
  MICRO_ENTREPRENEUR = 'MICRO',
  PRESTATION_ACCUEIL_JEUNE_ENFANT = 'PAJE',
  PRIME_ACTIVITE = 'PA',
  PENSION_INVALIDITE = 'PI',
  PENSIONS_ALIMENTAIRES = 'PAF',
  RSA = 'RSA',
  SALAIRE = 'PAIE',
  TRAVAILLEUR_INDEPENDANT = 'INDP'
}